requirements.txt
Flask==2.3.0
google-cloud-texttospeech==2.17.2
google-cloud-speech==2.27.0

